♻️ Extract 350+ lines from massive useEffect into focused hooks
Major refactoring to break down the 565-line useEffect monster:

- Extract settings management into useSettings hook (~220 lines)
  - All settings initialization and synchronization
  - Theme, sound, LLM, clipboard settings
  - Keyboard shortcuts management

- Extract drag & drop handling into useFileDrop hook (78 lines)
  - File drop event handling
  - Audio file validation
  - Upload triggering

- Extract transcript events into useTranscriptEvents hook (150 lines)
  - transcript-created listener
  - processing-complete listener
  - recording-completed listener
  - Auto-copy/paste logic
  - Success sound playing
  - Performance metrics logging

Results:
- App.tsx reduced from 1682 → 921 lines (45% reduction)
- Massive useEffect reduced from 565 → ~120 lines (78% reduction)
- Much better separation of concerns
- Each hook is focused, testable, and reusable
